Modal fixes complete (for now)
